Fix GL object release on window exit
[ ] May break build
[ ] May break existing applications
[x] Bugfix
[ ] New Feature
[ ] Cleanup
[ ] Optimization
[ ] Documentation

+++ examples/eqPly/window.cpp
@@ -100,7 +100,7 @@ bool Window::configInitGL( const uint32_t initID )
 
 bool Window::configExitGL()
 {
-    if( _state )
+    if( _state && !_state->isShared( ))
         _state->deleteAll();
 
     delete _state;
